Use specific QIO read/write functions for int/real/imag
Instead of calling QIO functions that dispatch based upon
an argument indicating the type. That is, instead of
calling
 qio_channel_write_int(.... , size=4 bytes, signed=yes)
we call
 qio_channel_write_int32(...) or qio_channel_write_uint46(...)
directly.

This change makes sense because the Chapel compiler already knows
the relevant types and so the call to qio_channel_write_int would
always be inlined for 1 case anyway. It also happens to avoid an
problem when using the intel compiler.