Revise liveVariableAnalysis to keep a named variable alive to the end of its scope
Added the function addUsesForNamedLocals(), which sets the "use" bit for each symbol whose
declaration scope contains the last statement in the current basic block.  This makes the
use bit true thoughout the scope of that variable (even before it is defined).  It is called
in the main basic block loop in liveVariableAnalysis.

At 104 regressions > 35 run-time > 20 non-leak.