Move tracking control entirely into chpl-mem-desc.
Move the control over which memory descriptors are tracked (as opposed
to the tracking itself) entirely into chpl-mem-desc.  Specifically, do
the check against the CHPL_RT_MEMTRACK_ALL_MDS environment variable from
chplmemtrack.c to chpl-mem-desc.c, next to the retrieval of the "track"
indicator from the table.

While here, add a note to the effect that we can imagine in the future
wanting finer control over which memory descriptors are tracked and
which are not., and this should be where we build that.  At present we
can only either track the ones the table says to (statically), or track
them all if the env var is set.