Fix chpldoc directory skipif for cygwin32 testing.
The test system wants to check this file out with executable bits, so instead
of fiddling with that, just use the executable skipif file format.

This is what the permissions look like after a fresh clone on the cygwin32 test
system using the simple .skipif file format:

$ ls -l $CHPL_HOME/test/chpldoc.skipif
-rw-rwxr--+ 1 chapelu nologon 87 May  4 12:35 $CHPL_HOME/test/chpldoc.skipif

The chpldoc.skipif now has 0755 permissions, so it should work as expected.