Default to not track mem desc types not directly related to Chapel code.
Change the default memtrack status for runtime-oriented memory desc
types that only have a loose relationship (especially an asynchronous
relationship) with Chapel program constructs.  Basically these are low
level buffers, flags, remote fork argument copies, and other comm info
blocks in the comm layer implementation, and task descriptors and other
meta-information having to do with task management in the tasking layer
implementation.  For the most part these are allocated synchronously as
execution encounters the corresponding Chapel program constructs (remote
communication events and tasking statements), but they may be freed
asynchronously with the end of execution of those constructs.

This improves the usefulness of memory tracking reporting, especially
when that is done in the middle of execution rather than at the end.  In
turn this reduces variability in memory tracking in nightly testing,
helping to resolve https://chapel.atlassian.net/browse/CHAPEL-8.

There was a choice here between continuing to track such memory descs
and just not reporting them, or not tracking them in the first place.
Tracking but not reporting would allow them to continue to affect the
current memory allocated and high-water mark values, but would create
(more of) a mismatch between the current memory allocated value and the
total of the by-type memory report, since they wouldn't appear in the
latter.  I chose instead to not track them at all, which makes the
current and high-water memory allocation values less accurate (because
the memory is still allocated even though it's not tracked) but removes
the tracking overhead.  It would not be much work to change to tracking
but not reporting if this is desired instead.