Adjust insertion of autoCopy calls in parallel to just begin, cobegin and coforall blocks.
Task functions that do not contain a _downEndCount call were causing a compile-time error
in symbol.cpp.  I backed off from adding autoCopies to all task blocks and used just
cobegin and coforall blocks.  The error persisted, so I changed
insertBeforeReturnAfterLabel to accommodate this.

Some compilation errors persist in case the search for a _downEndCount call runs outside
the enclosing block.  So this is a partial commit.

Down to 74 regressions > 27 run-time > 8 non-leak.