Fix new compile-time bugs.
There was an assertion failure in iterator.cpp.  Live variable analysis was revised to
include more variables.  Some of the new ones are reference variables, so we do not need
to convert them to reference typed fields in the iterator class.  I removed the assert and
replaced it with a conditional, so we perform the type conversion only when needed.

The code added to parallel.cpp to insert autoCopy and autoDestroy calls around cobegins
and coforalls caused an assertion to fire on line 2237 of symbol.cpp.  I further modified the
routine FnSymbol::insertBeforeDownEndCount() to tolerate the case where the call
expression being tested to see if it is a call to _downEndCount isn't a call at all but a
primitive instead.

At 55 regressions > 26 run-time > 8 non-leak.