Fix C Macro Bugs
* Enables --llvm to work with GMP
* fixes test cases just added to do with hex floats

Changed handleMacro to return void and instead a caller wishing to find the
value it added should just query the layer value table.  Fixed the case when a
macro is a C declaration already existing in the layered value table - now we
add an entry to the table from the macro name to the C declaration. In order to
do that, needed to add a new addGlobalCDecl that takes in a name to use
(instead of the name in a C NamedDecl).

test/extern/ferguson/externblock and test/modules/standard/gmp/ferguson
both pass with and without --llvm