Leak the origRootLocale.
At least the block distribution attempts to access the origRootLocale after it has been
deleted.  This is due to a mismatch between the module initialization order and the
assumption that shared objects will remain live until after their last use.

Locale objects are class objects, so their MM is under program (as opposed to compiler)
control.  The block distribution (for example) stores an array of locales (targetLocales),
but the locales themselves are stored elsewhere.  When running on a single locale, the
origRootLocale is never replaced.

It would seem that the distribution ProblemDist created in the user code should be deleted
before the origRootLocale.  This might go back to a problem in how the autoDestroyGlobals
is populated.

Anyway, the workaround (of not deleting origRootLocale) allows this and two other test
cases in the same directory to pass.