Remove SSCA2_mem performance test
This test has been timing out under --no-local performance testing which caused
us to look at it. It was originally added to track memory usage and leaks of
SSCA2 as it was being developed. It was not really meant to track memory
changes as a result of compiler/module changes. Since SSCA2 isn't being
actively developed, and nobody looks at these graphs, we're removing the test.

This removes the test and graphs.