Fix substr.chpl.
Revised the test to accommodate the new string implementation.

After renaming string.substring as string.this, he syntax <string-literal> '[]' does not
work, so I just stubbed out those tests.

c_string.this is not implemented, but c_string.substring() is, so I reverted those test to
call the .substring() method.

This test still fails, because of the open question regarding the semantics of
.this(i:integral) when the index falls outside of the given string (e.g. -1).