Fix bug in using inherited destructors
I discovered that if a class used its parent destructor, we'd encounter an
internal assertion error, because for some reason we didn't expect to insert
coercion temps.  This was a really dumb assertion, made from pure laziness, so
I removed it and wrapped the code in a block that could be removed so that what
the assertion was worried about didn't come to pass either.

Add a test to ensure this case will continue to pass.