Consolidate printing memTable records by type with printing them all.
##DETAILS

###modules/internal/MemTracking.chpl
###runtime/src/chplmemtrack.c
Renamed dumpMemLeaks as memLeaksByDesc.

###runtime/include/chplmemtrack.h
Updated interfaces to chpl_printMemAllocs and added chpl_printMemAllocsByDesc.
Added comments to distinuish the essential interface from those entry points intended for
debugging.

###runtime/src/chplmemtrack.c
Renamed chpl_dumpMemAllocs() as chpl_printMemAllocByDesc(), and reworked it to use
chpl_printMemAllocs() as its implementation.
Reworked chpl_printMemAllocs() to take a description argument, and filter which memTable
entries are printed using this argument.
Used the "if () continue;" idiom to filter out records that are not of interest.
Now pass in memThreshold to chpl_printMemAllocs, so it will respect the threshold when
printing out memTable entries.  This is probably more useful than ignoring it.

###test/execflags/shannon/configs/help/basehelp.txt
###test/execflags/shannon/help.good
Updated .good files to contain the new exec option.