Add tolerance to fileIO primer
This test fails under knc because of a floating point mismatch. We calculate
the values of an array in memory, write it out to a file, and then create a new
array from that file. However, we only write out 2 digits of precision to the
file so the values we read back in might not be identical. To compensate for
this, a tolerance is being added.

The particular issue was that when we calculate (1+7/10) intel was turning that
into (1+7*.1) which has a slightly different value than the literal 1.7 that we
were writing out to a file. It's likely this could have caused problems with
any configuration that enables replacing division with multiplication of a
reciprocal, but we only saw it on knc likely because of knc's different cost
model.

See JIRA issue 51 for more information