Suppress fasta-lines perf runs on machines with < 8 GB or ram
This test concats a bunch of strings and since it ends up leaking all of them
it requires ~6 GB of memory. Until we have strings as records, it makes sense
to just suppress this test since we're expecting it to fail.

The suppressif just checks if perf testing is on and if the system has < 8GB of
ram. Tested on linux64 and darwin.

Suppressif based off of modules/standard/memory/countMemory/countMemory.prediff