Quiet domainMaps regression on 32 bit platforms
This test leaks different amounts of memory for different configurations. The
current cases that we don't handle correctly are linux32 and cygwin32. This just
adds linux32 and cygwin32 platform specific .good files.

Long term we expect this test to have zero leaks and in the meantime we just
have different .good files for the configurations that leak different amounts.

I considered skipping this test for 32 bit platforms, or adding a supressif but
this was easier, and will allow us to track if things get worse on 32 bit for
some reason.

See JIRA issue 45 (https://chapel.atlassian.net/browse/CHAPEL-45)