Silence Coverity issue 1298873
I highly doubt that using our Vec class was why Coverity was complaining about
new_Expr (though the switch is a good thing to do anyways).  Instead, I believe
the complaint to be due to not understanding our calls to new_Expr() - namely,
our use should never result in certain characters being encountered first and
so it ought to be safe to look at the top of the stack when those characters
are encountered.  Such failure modes ought to be detected by parsing, or only
introduced by developers in later passes - to that end I'm inserting some
asserts here to validate that assumption.

If someone encounters these asserts, either they are doing something very wrong
or the designer of new_Expr did something wrong. In either case, we would obtain
undefined behavior from the stack otherwise - instead of a mysterious or hidden
error, it will now be out in the open.  If I encounter any errors through
testing because of these asserts, I'll resolve them.