Wrap some launcher errors to get better test reporting
We have two common errors that come from our launchers. One is an aprun missing
output file (Jira 17) and the other is a slurm expired credential (Jira 18).
The first first shows up as an error matching output and the second shows up as
a time out. This is just to make it obvious from emails what the actual error
is without having to go to the log.

Ideally we'd like to be able to get rid of these errors, but I have no idea
what the root cause it yet.