Remove backslashes from commands used to build target executables
The backslashes that were used in Makefile.exe made the Makefile
easier to read, but made the commands used to compile/link the
generated code harder to read.  Here I remove them to optimize
for what I think is the more common case.

Another approach that we could take to improving the Makefile's
readability is to combine some of the variables into other
variables to shorten the overall command-line, but I didn't
take that on here.