Do not submit an email address when uploading a build to Coverity.
It seems like Coverity sends only a "heartbeat" email to that address.
See an example below.

Currently we do not wish to receive such an email.
When needed, we can check Jenkins and Coverity instead.

So I am trying this change. If Coverity complains about the lack
of the 'email' field, we can revert this.

Sample email:

Date: .....
From: <scan-admin@coverity.com>
To: <the email address provided>
Subject: Coverity Scan: Analysis completed for chapel

    Your request for analysis of chapel has been completed successfully.
    The results are available at https://scan.coverity.com/projects/1222

    Analysis Summary:
       New defects found: 0
       Defects eliminated: 3