Fix failing sync future
This test verifies the behavior of sync variables of sync variables.  Both the
test .good file and I believe that this is not behavior we should allow.  The
failure mode is no longer going to be consistent, though, because syncs now get
freed during destruction.  Specifically, the sync clean causes readFE's attempts
to clean up after the temporaries it makes when reading from the value of a sync
variable to actually clean up the sync variable temporaries.  Since sync
variables don't copy themselves, cleaning up the "copy" is cleaning up the
result as well, which is why after the sync cleanup change we get garbage
instead of the sync's stored value.  Since we can't rely on a specific value for
the .bad file, I am removing it.  Skipping normalization for syncs may make this
future stable again.