Fix the gasnet regressions caused by closing sync and single leaks
These tests were encountering errors similar to what I fixed in the original
pull request, and the fix was the same.  Basically, a sync would sometimes go
out of scope and be destroyed before we were done using it.  This patch inserts
sync blocks around the usage of the syncs to prevent this error.