Replace use of our Vec class with use of std::vector in codegen of ModuleSymbols
Coverity was complaining about sending the Vec's n field to qsort because it
was possible the n field was a null pointer.  Since the Vec in question is a
local variable, I switched it to a std::vector and sent it to std::sort instead
of qsort (because the internet told me it that was a better sort to use
anyways).  Using vector's begin() and end() iterators also gets rid of the
complaint Coverity had, because the iterators won't dereference if there is
nothing in the vector.