Remove --no-local perf variants of lulesh and suppress for cce
We now have --no-local performance testing running against the entire
performance suite,  so it's redundant and unnecessary to do no-local testing
normally. This does mean we "lose" the history for the no-local runs, but we're
not too worried about that. The old data will still be in the .dat file it just
won't be graphed nightly.

There is also a cce bug (that will be fixed with 8.4) that causes the local
versions of this test to fail. The --no-local cases passed, so we couldn't
suppress this test with those still there. Since this removes the --no-local
variations, we can now suppress the failing --local versions.