Initial attempt at fixing coverity error 1298875
The fix that was applied because of 1298874 (convert to vector and switch the
sort used) doesn't apply as easily here.  Something about adding the FnSymbol
obtained from a forv_Vec iteration doesn't play nicely with a similar comparison
function - what is passed is the symbol pointer itself, not a const pointer to
the symbol pointer.  I need to do other, more pressing work, so save this state
for now.