Move distribution destruction to domain destructors.
There was an ordering problem in the destructor for distributed domains: When
the distribution reference count went to zero in destroyDom(), the distribution owned
by that domain was destroyed immediately.  But then the destructor for that same
domain uses the distribution (yes, the one that was just destroyed) to iterate
across locales and destroy the LocDoms it owns.

My solution was to push the destruction of the distribution into the individual
domain destructors.  This destruction explicitly follows the destruction of the
local doms, so the ordering problem is removed.

This change causes 38 read-after-free errors reported by valgrind to be cured.

TODO: This distributes the distribution destruction code to individual domains,
because that is easiest to get right.  Later, we can make this a more factored
solution by having individual domain destructors call a common cleanup utility
defined in BaseDom.