Move deletion of distributed domain to individual array destructors.
There was a use-after-free error reported by valgrind in BaseArr.destroyArr(), due to the
early call to the distributed domain destructor.  The domain is used in the array's
destructor to delete the elements of locArr, so it must remain live until after this.  By
moving the call to the distributed domain's destructor into individual distributed array
destructors, we can control the order of these actions and make the dom destruction occur
later.

TODO: We may need to add domain destructor calls for other distribution types.