Construct, not copy, records into arg bundles.
We used to copy records into the task functions' argument bundles
using bit copy. The exception was for 'begin' task functions,
where we would copy-construct the record field in the argument bundle
then destruct it within the task function.

This does not work with records with non-trivial constructors/destructors.
So this change extend copy-construction+destruction to all task functions.

We retain the prior situation, i.e. avoid constructing/destructing
for cobegins/coforalls, in the following cases:

* For reference-counted types, whose constructors/destructors only
manipulate the reference counter.

* For the first argument of an 'on' function, which is a
chpl_localeID_t dummy.

* When the record is known to be a POD. Currently this is indicated
with FLAG_COMPILER_GENERATED. See also markPODtypes().

In the last two cases, there is a danger of getting into an infinite
loop when a destructor has an 'on' clause. I added a comment to clarify.

I had to elaborate on insertBeforeDownEndCount(autoDestroyCall)
to handle these:

* In some cases the task function does not have an end count.

* In some cases the _downEndCount call is nested in one or more
PRIM_BLOCK_LOCAL blocks. This case seems to be specific to
being in the middle of the parallel() pass.