Rename FLAG_PRIVATE to FLAG_LOCALE_PRIVATE
In working to add the keywords public and private to the compiler, I discovered
that we already had a flag named private, but that it was very much intended to
represent locale-private.  To avoid confusion and simplify things, I renamed
FLAG_PRIVATE to FLAG_LOCALE_PRIVATE and updated all uses of it to reflect the
change made.  This included updating 5 internal module uses and 5 testing
system uses.