Bump ref count on adomain when creating a DefaultOpaqueArr.
The test case classes/bradc/arrayInClass/genericArrayInClass-otharrs was failing because
the implementation of chpl_incRefCountsForDomainsInArrayEltTypes(type eltType) creates a
realio-trulio DefaultOpaqueArray as the value of the variable ev if the element type is
"opaque".  Creating a DefaultOpaqueArray with the same DefaultOpaqueDom as another such
array causes the underlying DefaultSparseDom (==DefaultOpaqueDom.adomain) to be shared.
If we don't bump the reference count at that point, the shared adomain will be reclaimed
prematurely.