Default to fifo with pgi compiler
We're running into some weird build issues with pgi right now because of a .S
file in qthreads source. This is preventing the entire module from building so
we're swapping to use fifo until we figure out a real solution.

Updating the docs to mention when fifo is the default in case we never remember
to swap back or fifo ends up being a better option for pgi.