Put back in the zero-initialization of array fields in records and classes.
The zero-initialization that I disabled in commit cd97adf is necessary just for the case
that there is an array field in a class or record.  It is also necessary only because our
current implementation of field-initialziation in constructors confuses assignment with
initialization.

I put the workaround back in with tighter constraints on when it actually fires, and added
a long comment to aid in getting rid of the workaround after constructors are reworked to
use initialization (in the precise sense) for field-initialization.