Failed attempt to handle arrayInClassRecord valgrind failure.
Worked around the failure instead by adding a suppressif file.

The failure has to do with an array field initializer that is a forall expression.  The
forall expression is promoted to whole-array assignment that refers to the
yet-to-be-initialized array as an argument to the parloopexpr().  The uninitialized array
becomes the "this" field in the iterator record describing the iteration to be performed.
At the point in runtime when the iterator starts trying to reference elements of the
array, the reference to the array._value object has never actually been initialized.

This test passes without valgrind, but I assert that is purely by accident.  If, for
example, slots in a new stack frame were initialized to garbage rather than just
containing leftovers from the previous invocation, I suspect that the program would crash
rather than running successfully.