Remove a pair of lines that should, if my analysis is correct, do nothing
In scopeResolve's lookup, at one point we check if the scope we have just been
examining is in fact the scope of the surrounding module and if it is, we send
the scope of the surrounding module through lookup again.  Since looking at
scope added it to the list of already visited scopes, when we go into this
function again with the new value of scope being verified to be ourselves again
we return immediately.  This seems pointless, so I'm removing it.  Let's see
if anything breaks.