Remove code to give temporaries the same scope as target variables.
This was a workaround that probably because unnecessary after the rule for
inserting autodestroys was revised to place them at the end of the enclosing
scope.  Before, some variables were getting reclaimed after they were no longer
used.  Making them last as long as their declaration scope seems to be long
enough....

The workaround was causing a runtime failure in
  studies/amr/advection/amr/AMR_AdvectionCTUdriver
whieh is now resolved.