Suppress test_glob --baseline testing for the time being
Investigating yesterday, I verified that the test_glob failure
is an AMM-related issue due to double-freeing of a domain(string).
Tom verified that on the AMM branch it works as expected.  For
that reason, suppress this test for the time being and when AMM
is incorporated, we should be good to go.