Update the return symbol in eviscerated iterator functions.
[Open rant]
A long debugging session (5+ hours) followed by a very simple update.  This is
very strong use-case data that one should always either:
  * Not stash summary values extracted from data structures that may change
    and recompute/re-extract the desired value when needed; or
  * Encapsulate both in an abstract data type that will maintain the invariant
    relationship between the stashed value and the underlying data.
[Close rant]

In this case, rebuildIterator() replaced the return symbol with a different one
and failed to update the stashed retSymbol field.  Downstream decisions in
OwnershipFlowAnalysis depending on getReturnSymbol() returning the right answer
were fooled because getReturnSymbol() returns the stashed value without checking
for consistency between the stashed value and the actual value.

The manifestation of the problem was a memory leak in spectralnorm (and possibly
other programs) because getReturnSymbol() was used to determine if the symbol
being considered was the return value variable.  In that case, autoCopies are
turned off, because the return value variable (RVV) is assumed to already own
its contents.

Since the symbol did not match the RVV, an extra copy was inserted.  That left
the ref count of the referenced array at a value one greater than the actual
number of references present.  That in turn prevented the array and its contents
from ever being cleaned up.