Clean up NetworkAtomics line and file info stuff
We have an "insert line file info" pragma which automatically adds our line and
file info arguments to definition of exported functions and to all the call
sites of those functions.

This macro is used in other places in the modules that call out to the runtime,
but it wasn't being used for network atomics. Instead a manual version of this
was done where the line and file arguments were explicit in the exported
function definition and at every call site. This cleans that up to just use the
pragmas on the exported functions.

This is just meant as a clean up. I started going through the network atomics
code to investigated a bug with ugni network atomics that is currently the
biggest source of testing noise and thought this would make the module more
readable.