Resolve the valgrind error shown by genericArrayInClass-otharrs-inline-iterators.
This is a continuation of the effort to port the valgrind fix related to ForLoop
lowering from the string-as-rec branch to master.

Added the getIteratorInfoForIterator() function, mostly document the obscure way
in which we obtain the type of the underlying iterator in a ForLoop statement.

Rewrote the code that populates the map used in copying parts of a loop
construct into the various zip?() functions, hasMore(), getValue(), etc. to
clarify what it is actually doing and why.

Added code that calls down to the corresponding function in the underlying
iterator, to impelement the components of a forLoop.

Replaced "ii->foo->_this" with "my_this" where possible.

Renamed "flag" as "beforeYield", to make it more self-documenting.

Renamed getMoreBody to getValueBody inside the getValue() function, for
consistency.

Removed the genericArrayInClass-otharrs-inline-iterators test case, as it is no
longer needed.

Removed genericArrayInClass-otharrs.compopts.  It used to contain
--no-inline-iterators as a workaround for the bug that this commit fixes.  With
the fix, the workaround is no longer needed.