Fixed non-determinism in 21-capture-in-cobegin et al. JIRA #63: https://chapel.atlassian.net/browse/CHAPEL-63
THE ISSUE

In the following code

```
    var x = 5;
    ....
    coforall .... do
      writeln(x);
```

we want all iterations of coforall to print the value of 'x'
that 'x' had immediately before entering 'coforall'.
As a result, all iterations are guaranteed to print the same.

Prior to this change, our implementation was capturing
the value of 'x' at the time each coforall task was created,
i.e. when initializing each argument bundle.
So it was possible that 'x' changes between initializations
of the first and second bundle, violating our desired property.

Here are a couple of fleshed-out examples:

```
    var x = 5;
    begin with (ref x) { x = 6; }
    coforall ... do writeln(x);
```

```
    var x = 5;
    proc update() { x = 6; }
    cobegin { /*Task1*/writeln(x); /*Task2*/update(); }
```

THE FIX

For each variable to be captured, I create a temp and initialize
it from that variable right before the start of the corresponding
task construct. That temp is used instead of the original variable
when initializing the argument bundles.

WHAT'S UP WITH RECORDS?

Right now the blank task intent for records is incorrectly implemented
as 'const in'. Since this is incorrect, I explicitly exclude records
from The Fix above. Therefore the records continue to be captured at
task creation time. I adjusted the tests not to test records, because
the behavior remains nondeterministic.

WHAT'S WITH STRINGS AND UNIONS?

What I wrote about records above should apply to strings and unions.
I do not know for sure whether it does. For now I exclude strings and
unions from testing as well, because most likely their behavior in
this regard currently is also most likely nondeterministic.